+++ apps/admin/.graphql-let.yml
@@ -4,6 +4,7 @@ documents:
 plugins:
   - typescript-operations
   - typescript-react-apollo
+  # - typescript-urql
 cacheDir: __generated__
 config:
   avoidOptionals: false
