+++ apps/admin/package.json
@@ -120,6 +120,7 @@
     "@graphql-codegen/typescript-operations": "^2.5.6",
     "@graphql-codegen/typescript-react-apollo": "^3.3.6",
     "@graphql-codegen/typescript-resolvers": "^2.7.6",
+    "@graphql-codegen/typescript-urql": "^3.7.3",
     "@types/gtag.js": "^0.0.12",
     "@types/jest": "^29.5.3",
     "@types/multer": "^1.4.5",
