+++ apps/admin/src/app/post/[postId]/loading.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+const Loader = () => {
+  return <h1 style={{ background: "red" }}>Loading....</h1>;
+};
+
+export default Loader;
