+++ apps/admin/src/features/domain-mapping/index.ts
@@ -1,0 +1,1 @@
+export { Feature } from "./feature";
